display modal from agGridClick
