+++ src/app/rickAndMorty/components/rick-and-morty-characters-grid/rick-and-morty-characters-grid.component.html
@@ -13,4 +13,5 @@
 </div>
 <div class="flex justify-center mt-2">
     <app-add-new-character-button />
+    <app-update-rick-and-morty-character-modal></app-update-rick-and-morty-character-modal>
 </div>
