+++ src/app/rickAndMorty/components/update-rick-and-morty-character-modal/update-rick-and-morty-character-modal.component.html
@@ -1,0 +1,53 @@
+
+<!-- Main modal -->
+<div id="update-character-modal" tabindex="-1" aria-hidden="true"
+    class="hidden overflow-y-auto overflow-x-hidden fixed top-0 right-0 left-0 z-50 justify-center items-center w-full md:inset-0 h-[calc(100%-1rem)] max-h-full">
+    <div class="relative p-4 w-full max-w-md max-h-full">
+        <!-- Modal content -->
+        <div class="relative bg-white rounded-lg shadow">
+            <!-- Modal header -->
+            <div class="flex items-center justify-between p-4 md:p-5 border-b rounded-t ">
+                <h3 class="text-lg font-semibold text-gray-900 ">
+                    Update character
+                </h3>
+                <button type="button"
+                    class="text-gray-400 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm w-8 h-8 ms-auto inline-flex justify-center items-center"
+                    data-modal-toggle="update-character-modal">
+                    <svg class="w-3 h-3" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none"
+                        viewBox="0 0 14 14">
+                        <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+                            d="m1 1 6 6m0 0 6 6M7 7l6-6M7 7l-6 6" />
+                    </svg>
+                    <span class="sr-only">Close modal</span>
+                </button>
+            </div>
+
+
+            <!-- Modal body -->
+            <form class="p-4 md:p-5" name="addCharacterForm" [formGroup]="updateCharacterForm" (ngSubmit)="onSubmit()">
+                <div class="grid gap-4 mb-4 grid-cols-2">
+                    <div class="col-span-2 sm:col-span-1">
+                        <label for="lifeStatus" class="block mb-2 text-sm font-medium text-gray-900 ">Life
+                            Status</label>
+                        <select id="lifeStatus" name="lifeStatus" formControlName="lifeStatus"
+                            class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2.5">
+                            <option [ngValue]=0>Alive</option>
+                            <option [ngValue]=1>Dead</option>
+                            <option [ngValue]=2>Unknown</option>
+                        </select>
+                    </div>
+                </div>
+                <button type="submit" data-modal-toggle="update-character-modal" [disabled]="!updateCharacterForm.valid"
+                    class="disabled:opacity-30 disabled:cursor-not-allowed text-white inline-flex items-center bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center">
+                    <svg class="me-1 -ms-1 w-5 h-5" fill="currentColor" viewBox="0 0 20 20"
+                        xmlns="http://www.w3.org/2000/svg">
+                        <path fill-rule="evenodd"
+                            d="M10 5a1 1 0 011 1v3h3a1 1 0 110 2h-3v3a1 1 0 11-2 0v-3H6a1 1 0 110-2h3V6a1 1 0 011-1z"
+                            clip-rule="evenodd"></path>
+                    </svg>
+                    Add new character
+                </button>
+            </form>
+        </div>
+    </div>
+</div>
